+++ packages/orq-rc/RELEASES.md
@@ -7669,3 +7669,13 @@ Based on:
 - [typescript v4.6.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v4.6.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.6.0-rc.5 - packages/orq-rc
+
+## 2026-03-12 12:09:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.752.0 (2.859.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.6.0-rc.6] packages/orq-rc
+### Releases
+- [NPM v4.6.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.6.0-rc.6 - packages/orq-rc